+++ src/propertyFields/listPicker/PropertyFieldListMultiPickerHost.tsx
@@ -61,6 +61,8 @@ export default class PropertyFieldListMultiPickerHost extends React.Component<IP
   private loadLists(): void {
     // Builds the SharePoint List service
     const listService: SPListPickerService = new SPListPickerService(this.props, this.props.context);
+    const listsToExclude: string[] = this.props.listsToExclude || [];
+
     this.options = [];
     // Gets the libs
     listService.getLibs().then((response: ISPLists) => {
@@ -76,12 +78,15 @@ export default class PropertyFieldListMultiPickerHost extends React.Component<IP
           isSelected = true;
           this.state.selectedKeys.push(list.Id);
         }
-        // Add the option to the list
-        this.options.push({
-          key: list.Id,
-          text: list.Title,
-          checked: isSelected
-        });
+
+        // Add the option to the list if not inside the 'listsToExclude' array
+        if (listsToExclude.indexOf(list.Title) === -1 && listsToExclude.indexOf(list.Id) === -1) {
+          this.options.push({
+            key: list.Id,
+            text: list.Title,
+            checked: isSelected
+          });
+        }
       });
       this.loaded = true;
       this.setState({