+++ src/webparts/propertyControlsTest/PropertyControlsTestWebPart.ts
@@ -7,7 +7,7 @@ import {
   BaseClientSideWebPart,
   IPropertyPaneConfiguration,
   PropertyPaneTextField,
-  PropertyPaneToggle,
+  PropertyPaneToggle
 } from '@microsoft/sp-webpart-base';
 import { PropertyFieldCodeEditor,PropertyFieldCodeEditorLanguages } from '../../PropertyFieldCodeEditor';
 import * as strings from 'PropertyControlsTestWebPartStrings';
@@ -160,6 +160,8 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   enableSorting: true,
                   disableItemDeletion: false,
                   disableItemCreation: false,
+                  panelClassName: "MyAwesomePanelClassName",
+                  tableClassName: "MyAwesomeTableClassName",
                   fields: [
                     {
                       id: "Title",
@@ -343,7 +345,8 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   },
                   deferredValidationTime: 0,
                   key: 'listPickerFieldId',
-                  webAbsoluteUrl: this.properties.siteUrl || this.context.pageContext.web.absoluteUrl
+                  webAbsoluteUrl: this.properties.siteUrl || this.context.pageContext.web.absoluteUrl,
+                  listsToExclude: ["cdn"]
                 }),
                 PropertyFieldListPicker('multiList', {
                   label: 'Select multiple lists',
@@ -362,7 +365,8 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   onGetErrorMessage: null,
                   deferredValidationTime: 0,
                   key: 'multiListPickerFieldId',
-                  webAbsoluteUrl: this.properties.siteUrl || this.context.pageContext.web.absoluteUrl
+                  webAbsoluteUrl: this.properties.siteUrl || this.context.pageContext.web.absoluteUrl,
+                  listsToExclude: ["cdn"]
                 }),
                 PropertyFieldDateTimePicker('datetime', {
                   label: 'Select the date and time',
@@ -376,7 +380,8 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   properties: this.properties,
                   onGetErrorMessage: null,
                   deferredValidationTime: 0,
-                  key: 'dateTimeFieldId'
+                  key: 'dateTimeFieldId',
+                  showLabels: false
                 }),
                 PropertyPaneToggle("isColorFieldVisible", {
                   label: "Color Field Visible",
@@ -474,15 +479,24 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   options: [{
                     key: 'iOS',
                     text: 'iOS',
-                    checked: this.properties.choiceGroupWithCalloutValue === 'iOS'
+                    checked: this.properties.choiceGroupWithCalloutValue === 'iOS',
+                    iconProps: {
+                      officeFabricIconFontName: 'CheckMark'
+                    }
                   }, {
                     key: 'Android',
                     text: 'Android',
-                    checked: this.properties.choiceGroupWithCalloutValue === 'Android'
+                    checked: this.properties.choiceGroupWithCalloutValue === 'Android',
+                    iconProps: {
+                      officeFabricIconFontName: 'CheckMark'
+                    }
                   }, {
                     key: 'Other',
                     text: 'Other',
-                    checked: this.properties.choiceGroupWithCalloutValue === 'Other'
+                    checked: this.properties.choiceGroupWithCalloutValue === 'Other',
+                    iconProps: {
+                      officeFabricIconFontName: 'CheckMark'
+                    }
                   }]
                 }),
                 PropertyFieldButtonWithCallout('fakeProperty', {